Speed it up, then slow it down.
This optimizes the electronics simulation to use an FFT to convolve the
electronics delta function pulse (instead of using a direct convolution).
It also under-samples the electron distribution by a factor so that extra
fluctuations are less than the ADC digitization step.  These amount to a
large speed up.  The optimization are offset by simulating the noise in
the entire digitization range and expanding the digitization range from 1.5
to 3 milliseconds.  The net result is that the simulation is about 30%
faster.