Add a slightly more realistic simulation of the induction wire signal.
The induced charge on an induction wire is now simulated a little bit more
carefully.  See Mathieson for a complete treatment of charge induced on a
wire.  For now, I'm "averaging" over all charges and treating the amount of
induced charge as a Gaussian in time that is tied to zero at 3*mm.  Check
the code for the actual approximation, it's pretty crude, but not quite as
crude as the last one.