Significant changes to make elecSim more predictive for the TPC charge.
This commit includes the following changes:

1) Use output from NEST in DETSIM if it's available to calculate the number
of electrons generated.

2) Improve the model of the induced charge on wire.  This uses
Shockley-Ramo to estimate the shape of the induced current by approximating
the electron path and the weighting field.  The electrons are assumed to
drift with constant velocity and in a straight line at a particular impact
parameter relative to the wire.  The weighting field is assumed to be the
same as a single wire, but with the condition that the field is zero at the
grid planes.  The model incorrectly estimates the magnitude of the induced
current, but is useful for estimating the time evolution of the current.
The magnitude of the current is estimated separately.