Limit the maximum fluctuation due to digitizer noise.
The elecSim is (still) over estimating the amount of random interference in
the induction wire signal.  It probably affects the collection wires too,
but it's not as apparent.  The problem is a "wrong" noise model in that
most of the "noise" isn't noise, but interference from LRC components, or
external sources.  The interference is being modeled as a Gaussian noise,
which is overly pessimistic.