Mostly fix the normalization of the background noise.
Almost everyplace else, the normalization of the FFT is managed by taking
both the FFT and the inverse FFT (a lot of the constants cancel).  However,
the pink noise is added as a power spectrum and the normalization is
defined based on the convolution with the electronics response.  That
means it needs to account for all of the missing constants.  This commit
tries to collect all of the missing factors of 2, sqrt(2) and pi into one
spot.  The value was determined "empirically".