Add a scale factor for the overall power in the discreet noise peaks.
The discreet noise peaks are simulated using a random power (per peak) with
a random phase.  The power is chosen from a Gaussian centered on the
expected power.  The phase is uniform from 0 to 2pi.  The resulting
distribution only approximates the power in the noise spectrum, so there is
an additional tuning parameter added to fix the normalization.