Finish factoring to model before and after the shaping.
This fixes the model to more explicitly use charge before the asic, and
voltage after the asic.  The digitizer then translates the voltage into
digits.

Implements a consistent noise model based on data.  Noise is modeled as a
continuous spectrum plus discreet frequency bands.  A component of the
noise is added before the shaping to account for "shot" noise.

The FFT and inverse FFT have been explicitly broken out of the rest of the
code.  The normalization for both is now applied so that the FFT represents
the amplitude at each frequency.