Tweak the way the noise power is chosen.
This changes the way the noise amplitude is chosen.  The amplitude is chosen
randomly from a Gaussian centered on the measured noise, and with a width
determined from the data.  Previously, if a negative amplitude was chosen,
it was set to be zero.  Now, if a negative amplitude is chosen, a new value
is chosen until it is positive.